requirements.txt
-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
certifi==2021.5.30
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dj-database-url==0.5.0
django-cors-headers==3.7.0
django==3.2.5
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pillow==8.3.1
psycopg2-binary==2.9.1
psycopg2==2.9.1
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1; python_version >= '3.5'
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
whitenoise==5.2.0
